Fix failing dialogs after recent API change

+++ animation_nodes/nodes/subprogram/invoke_subprogram.py
@@ -196,7 +196,7 @@ class ChangeSubprogram(bpy.types.Operator):
             node = getNodeByIdentifier(self.nodeIdentifier)
             self.subprogram = node.subprogramIdentifier
         except: pass # when the old subprogram identifier doesn't exist
-        return context.window_manager.invoke_props_dialog(self, width = 400 * getDpiFactor())
+        return context.window_manager.invoke_props_dialog(self, width = int(400 * getDpiFactor()))
 
     def check(self, context):
         return True
